ai_module/requirements.txt
aim
matplotlib
numpy
Pillow
torch
torchvision
tqdm
psycopg2-binary